backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
python-dotenv==1.0.0
firebase-admin==6.2.0
pytest==7.4.0
gunicorn==21.2.0
flask-restful==0.3.10
bcrypt==4.0.1
PyJWT==2.8.0
